pg_dump: Add --extra-dependencies option
This patch introduces the new --extra-dependencies option to pg_dump.
This allows users to provide additional dependency hints for
reordering dumpable objects.  The goal is to overcome pg_upgrade
failures that occur when extensions have dependencies not fully
tracked in pg_depend.

The option accepts a comma-separated list of dependencies in the
format: "table-name#ref-table-name".

When this option is used, the data for the referred table
(ref-table-name) is dumped immediately after its definition is
created, utilizing the dumpTableData() infrastructure.  This ensures
that any subsequent object that depends on the data of the referred
table (and not just its schema) does not fail during restore.

For now, filter conditions are ignored during this immediate dump.
Once the data is dumped, the corresponding dump object (dobj.dump)
is set to DUMP_COMPONENT_NONE to prevent the data from being dumped
again during the regular flow.

Jeevan Chalke, per suggestion from Alvaro Herrera.